requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
Cython==0.29.33
deprecation==2.1.0
Django==3.2.16
django-admin-rangefilter==0.9.0
django-cors-headers==3.13.0
django-elasticsearch-dsl==7.2.0
django-environ==0.9.0
django-nine==0.2.7
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.14.0
elasticsearch-dsl==7.4.0
et-xmlfile==1.1.0
filelock==3.8.0
flower @ https://github.com/mher/flower/zipball/master
gunicorn==20.1.0
humanfriendly==10.0
humanize==4.5.0
idna==3.4
img2pdf==0.4.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
NoelOCR==0.0.8
ocrmypdf==14.0.2
openapi-codec==1.3.2
openpyxl==3.1.1
packaging==23.0
pdf2image==1.16.2
pdfminer.six==20221105
phonenumbers==8.13.4
pikepdf==7.0.0
Pillow==9.0.1
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyPDF2==3.0.1
pytesseract==0.3.8
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.4.2
reportlab==3.6.12
requests==2.28.1
requests-file==1.5.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tesseract==0.1.3
tldextract==3.4.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
Wand==0.6.7
wcwidth==0.2.6
zipp==3.11.0
